feat: refactor account components to use useAuth hook and streamline authentication logic
