+++ client/src/pages/account/Account.css
@@ -52,9 +52,3 @@
 .account-pass-main {
     width: 90%;
 }
-
-hr {
-    opacity: 0.7;
-    margin: 20px 0;
-    background: #1976d2;
-}
